[funding]Feign client 수정 - 조각, 결제 DTO 수정 - 결제 ENUM 생성 - feign client 수정 - 구매 조각 조회 sql문 수정 - x-member-uuid feign config 파일 추가
#35

+++ src/main/java/com/pieceofcake/fundingservice/participation/application/RedisService.java
@@ -6,9 +6,9 @@ import com.pieceofcake.fundingservice.funding.dto.in.UpdateFundingRequestDto;
 
 public interface RedisService {
     int getRemainingPieces(String fundingUuid);
-    double getPiecePrice(String fundingUuid);
+    long getPiecePrice(String fundingUuid);
     void setRemainingPieces(SetRedisFundingRequestDto setRedisFundingRequestDto);
     void deleteRemainingPieces(String fundingUuid);
-    Long decreaseRemainPieces(String fundingUuid, int quantity);
+    long decreaseRemainPieces(String fundingUuid, int quantity);
     boolean increaseRemainPieces(String fundingUuid, int quantity);
 }
